packages/embed-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embed-core"
version = "0.1.0"
description = "Abstract embedding providers and vector store interfaces"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
]

[project.optional-dependencies]
clip = [
    "open-clip-torch>=2.24",
    "Pillow>=10.3",
    "torch>=2.3",
    "torchvision>=0.18",
]
sentence-transformers = [
    "sentence-transformers>=3.0",
    "Pillow>=10.3",
]
bedrock = [
    "boto3>=1.34",
    "Pillow>=10.3",
]
openai = [
    "openai>=1.30",
]
pgvector = [
    "asyncpg>=0.29",
    "pgvector>=0.3",
]
pinecone = [
    "pinecone-client>=4.1",
]
opensearch = [
    "opensearch-py>=2.6",
    "aiohttp>=3.9",
]
all = [
    "embed-core[clip,sentence-transformers,bedrock,openai,pgvector,pinecone,opensearch]",
]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "mypy>=1.10",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/embed_core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.ruff]
line-length = 100
target-version = "py311"
